Don't use maxChunkSize if adaptive is false
The README says the maxChunkSize option is only used if adaptive = false https://lztc.qonq.gq
But this is not true; in the code there is a second copy of the "Limit to max chunk size" code, outside of the code block for adaptive mode. So either this should be removed to comply with the documentation, or the documentation should be updated; and in that case, the "Limit to max chunk size" code inside the adaptive block is unnecessary as the exact same code is ran again right after it.